Updated the MainController script
Separated into two parts
