+++ CSS551MP5_RayMichael/Assets/MainController_CamManipulation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 204286acd922f0a44a858ab023345812
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
